objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7138 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 9c00b54944f0b2f027d0438171924050364cc9aa

+++ trunk/xmipp/libraries/reconstruction/fourier_filter.cpp
@@ -256,7 +256,7 @@ void ProgFourierFilter::show()
     }
 }
 
-void ProgFourierFilter::processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+void ProgFourierFilter::processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
 {
     static bool first = true;
